move helper classes to separate files
