+++ src/main/resources/physiology/generators/circulation_hemodynamics.yml
@@ -50,7 +50,7 @@ inputs:
     - fromExp: "1.1889 + 0.025 * (#{BMI} - 21.7) + (if #{blood_pressure_controlled} != 0 then 0 else (0.0112 * #{age}))"
       to: R_sys
       variance: 0.25 # Allowable variance in input to avoid running simulation
-    - fromExp: "0.8413 + (if #{blood_pressure_controlled} != 0 then 0 else (if #{age} > 20 then 0.0175 * (#{age} - 20) else 0))"
+    - fromExp: "0.8413 + (if #{age} > 20 then 0.0175 * (#{age} - 20) else 0)"
       to: E_es_ao
       variance: 0.2 # Allowable variance in input to avoid running simulation
