+++ src/main/resources/synthea.properties
@@ -182,4 +182,4 @@ lifecycle.adherence.baseline = 0.05
 lifecycle.death_by_natural_causes = false
 
 # Use physiology simulations
-physiology.enabled = false
+physiology.enabled = true
